# screen_system/openapi_before.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "スクリーン原団 工程管理 v2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/export/inventory.csv": {
      "get": {
        "operationId": "admin_export_inventory_api_admin_export_inventory_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "pw",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pw",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Export Inventory"
      }
    },
    "/api/admin/export/orders.csv": {
      "get": {
        "operationId": "admin_export_orders_api_admin_export_orders_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "pw",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pw",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Export Orders"
      }
    },
    "/api/admin/fire-report": {
      "get": {
        "operationId": "admin_fire_report_api_admin_fire_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "title": "Month",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Fire Report"
      }
    },
    "/api/admin/fire-report.csv": {
      "get": {
        "operationId": "admin_fire_report_csv_api_admin_fire_report_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "title": "Month",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pw",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pw",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Fire Report Csv"
      }
    },
...
    "/api/admin/scans": {
      "get": {
        "operationId": "admin_scans_api_admin_scans_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "default": "",
              "title": "Stage",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Scans"
      }
    },
    "/api/admin/scans.csv": {
      "get": {
        "operationId": "admin_scans_csv_api_admin_scans_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "default": "",
              "title": "Stage",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pw",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pw",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Scans Csv"
      }
    },
    "/api/admin/scans/{scan_id}": {
      "delete": {
        "description": "スキャン履歴を1件削除する(誤スキャン・テスト記録の掃除用)。\n\n注意: 履歴を消しても order_items.current_stage は動かない。\n工程を戻したい場合は注文詳細の「工程で補正」を使うこと。",
        "operationId": "admin_scan_delete_api_admin_scans__scan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Scan Delete"
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "operationId": "admin_stats_api_admin_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Stats"
      }
    },
...
  }
}